implemented event consumed notification
Every time `Widget::handle_event` returns `EventResult::Consumed`, it
will trigger a call to `Widget::event_consumed` on the entire widget
tree, and pass info about the event and who consumed it. The act of
consuming an event is an event itself, and can be handled anywhere on
the widget tree. This way we can react to things happening elsewhere,
like a child getting clicked with the handler code on the parent.